+++ app/3d/lights.tsx
@@ -13,8 +13,7 @@ export default function DefaultDirectionalLight(props: DirectionalLightProps) {
 export function DirectionalLightWithShadows(props: DirectionalLightProps) {
     return <DefaultDirectionalLight
         castShadow
-        shadow-mapSize={[2048, 2048]}
-        shadow-normalBias={ 0.04 }
+        shadow-mapSize={[4096, 2048]}
         {...props}
     />
 }
