+++ app/routes/demos/03-debug-ui/index.tsx
@@ -1,8 +1,11 @@
+import { useEffect } from "react";
+
 import MainScene from "./main-scene";
 
 import type { Route } from "./+types/index";
 
 import { TestCanvas } from "~/3d/test-canvas";
+import { useSidePanel } from "~/stores/side-panel";
 
 export function meta(_: Route.MetaArgs) {
   return [
@@ -13,6 +16,12 @@ export function meta(_: Route.MetaArgs) {
 
 
 export default function DebugUI() {
+  const setIsOpen = useSidePanel((s) => s.setIsOpen);
+
+  useEffect(() => {
+    setIsOpen(true);
+  }, []);
+
   return(<TestCanvas>
     <MainScene />
   </TestCanvas>);
