+++ app/routes/demos/05-load-models/fox.tsx
@@ -2,14 +2,23 @@ import { useAnimations, useGLTF } from "@react-three/drei";
 import { useControls } from "leva";
 import { useEffect } from "react";
 
-import type { AnimationAction } from "three";
+import type { AnimationAction, Mesh } from "three";
 import type { MeshProps } from "~/types/types";
 
 const FOX_URL = "https://raw.githubusercontent.com/KhronosGroup/glTF-Sample-Assets/refs/heads/main/Models/Fox/glTF/Fox.gltf";
 
-export default function Fox(props: MeshProps) {
+export default function Fox({ castShadow, receiveShadow, ...props }: MeshProps) {
   const model = useGLTF(FOX_URL);
   const animations = useAnimations(model.animations, model.scene);
+
+  useEffect(() => {
+    model.scene.traverse((child) => {
+      if ((child as Mesh).isMesh) {
+        child.castShadow = castShadow ?? false;
+        child.receiveShadow = receiveShadow ?? false;
+      }
+    });
+  }, [model.scene, castShadow, receiveShadow]);
 
   const { animationName } = useControls("Fox", {
     animationName: { options: animations.names, label: "Animation" },
