+++ app/routes/demos/12-mini-game/player/use-player-controls.ts
@@ -1,18 +1,11 @@
 import { useKeyboardControls } from "@react-three/drei";
 import { useFrame } from "@react-three/fiber";
-import { RapierRigidBody, RigidBody, useRapier } from "@react-three/rapier";
-import { useCallback, useEffect, useRef, useState } from "react";
-import { Vector3 } from "three";
+import { RapierRigidBody, useRapier } from "@react-three/rapier";
+import { useCallback, useEffect, useRef, type RefObject } from "react";
 
-export default function Player() {
-  const playerRef = useRef<RapierRigidBody>(null);
+export function usePlayerControls(playerRef: RefObject<RapierRigidBody | null>) {
   const impulseRef = useRef({ x: 0, y: 0, z: 0 });
   const torqueRef = useRef({ x: 0, y: 0, z: 0 });
-  const cameraPositionRef = useRef(new Vector3(0, 0, 0));
-  const cameraTargetRef = useRef(new Vector3(0, 0, 0));
-
-  const [smoothedCameraPosition] = useState(() => new Vector3(10, 10, 10));
-  const [smoothedCameraTarget] = useState(() => new Vector3(0, 0, 0));
 
   const { rapier, world } = useRapier();
   const [subscribeKeys, getKeys] = useKeyboardControls();
@@ -42,7 +35,7 @@ export default function Player() {
     return () => unsubscribeKeys();
   }, []);
 
-  useFrame(({ camera }, delta) => {
+  useFrame((_, delta) => {
     if (!playerRef.current) return;
 
     const { forward, backward, leftward, rightward } = getKeys();
@@ -66,39 +59,5 @@ export default function Player() {
 
     playerRef.current.applyImpulse(impulse, true);
     playerRef.current.applyTorqueImpulse(torque, true);
-
-    const bodyPosition = playerRef.current.translation();
-    cameraPositionRef.current.copy(bodyPosition);
-    cameraPositionRef.current.y += 0.65;
-    cameraPositionRef.current.z += 2.25;
-
-    cameraTargetRef.current.copy(bodyPosition);
-    cameraTargetRef.current.y += 0.25;
-
-    smoothedCameraPosition.lerp(cameraPositionRef.current, 5 * delta);
-    smoothedCameraTarget.lerp(cameraTargetRef.current, 5 * delta);
-
-    camera.position.copy(smoothedCameraPosition);
-    camera.lookAt(smoothedCameraTarget);
-
   });
-
-  return <RigidBody
-    ref={playerRef}
-    colliders="ball"
-    canSleep={false}
-    position={[0, 1, 0]}
-    restitution={0.2}
-    friction={1}
-    linearDamping={0.5}
-    angularDamping={0.5}
-  >
-    <mesh castShadow>
-      <icosahedronGeometry args={[0.3, 1]} />
-      <meshStandardMaterial
-        color="mediumpurple"
-        flatShading
-      />
-    </mesh>
-  </RigidBody>;
 }
